+++ src/main/java/com/web/doitcommit/controller/HashtagCategoryController.java
@@ -1,9 +1,9 @@
 package com.web.doitcommit.controller;
 
 import com.web.doitcommit.dto.CMRespDto;
-import com.web.doitcommit.dto.popularTag.PoplarTagResDto;
-import com.web.doitcommit.dto.todo.TodoResDto;
-import com.web.doitcommit.service.popularTag.PopularTagService;
+import com.web.doitcommit.dto.hashtagCategory.HashtagCategoryResDto;
+import com.web.doitcommit.dto.hashtagCategory.PopularTagResDto;
+import com.web.doitcommit.service.hashtagCategory.HashtagCategoryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -22,16 +22,33 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
-public class PopularTagController {
+public class HashtagCategoryController {
 
-    private final PopularTagService popularTagService;
+    private final HashtagCategoryService hashtagCategoryService;
+
+    /**
+     * 해시태그 목록 조회
+     */
+    @Operation(summary = "해시태그 목록 조회 API", description = "해시태그 목록 조회을 조회한다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", content = @Content(array = @ArraySchema(schema = @Schema(implementation = HashtagCategoryResDto.class)))),
+            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(example = "{\"error\": \"Bad Request\"}"))),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(example = "{\"error\": \"Unauthorized\"}"))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(example = "{\"error\": \"Not Found\"}"))),
+            @ApiResponse(responseCode = "500",  content = @Content(schema = @Schema(example = "{\"error\": \"Internal Server Error\"}")))
+    })
+    @GetMapping("/tag")
+    public ResponseEntity<?> getTagList() {
+        List<HashtagCategoryResDto> hashtagCategoryResDto = hashtagCategoryService.getBoardTagList();
+        return new ResponseEntity<>(new CMRespDto<>(1, "태그 리스트 조회 성공", hashtagCategoryResDto),HttpStatus.OK);
+    }
 
     /**
      * 지정기간 동안 {tagId, tagName, count} 전체 인기 태그 리스트
      */
     @Operation(summary = "인기태그 리스트 조회 api", description = "지정기간 동안 인기태그 리스트를 조회한다. default : 7일")
     @ApiResponses({
-            @ApiResponse(responseCode = "200", content = @Content(array = @ArraySchema(schema = @Schema(implementation = PoplarTagResDto.class)))),
+            @ApiResponse(responseCode = "200", content = @Content(array = @ArraySchema(schema = @Schema(implementation = PopularTagResDto.class)))),
             @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(example = "{\"error\": \"Bad Request\"}"))),
             @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(example = "{\"error\": \"Unauthorized\"}"))),
             @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(example = "{\"error\": \"Not Found\"}"))),
@@ -39,7 +56,7 @@ public class PopularTagController {
     })
     @GetMapping("/popularTags")
     public ResponseEntity<?> getAllPopularTagList(@RequestParam(value = "period", required = false, defaultValue = "7") int period) {
-        List<PoplarTagResDto> popularTagResDtoList = popularTagService.getAllPopularTagListForPeriod(period);
+        List<PopularTagResDto> popularTagResDtoList = hashtagCategoryService.getAllPopularTagListForPeriod(period);
 
         return new ResponseEntity<>(new CMRespDto<>(
                 1, "지정된 기간동안의 인기태그 리스트 불러오기 성공", popularTagResDtoList), HttpStatus.OK);
@@ -72,7 +89,7 @@ public class PopularTagController {
     @Operation(summary = "7일동안 상위 8개의 인기태그 리스트 조회 api", description = "7일 동안의 상위 8개 인기태그를 조회한다. -> 7일동안 쌓이 데이터가 없으면 " +
             "전체 기간 상위 8개의 인기태그 리스트 조회")
     @ApiResponses({
-            @ApiResponse(responseCode = "200", content = @Content(array = @ArraySchema(schema = @Schema(implementation = PoplarTagResDto.class)))),
+            @ApiResponse(responseCode = "200", content = @Content(array = @ArraySchema(schema = @Schema(implementation = PopularTagResDto.class)))),
             @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(example = "{\"error\": \"Bad Request\"}"))),
             @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(example = "{\"error\": \"Unauthorized\"}"))),
             @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(example = "{\"error\": \"Not Found\"}"))),
@@ -80,12 +97,12 @@ public class PopularTagController {
     })
     @GetMapping("/popularTags/limit")
     public ResponseEntity<?> getLimitPopularTagListFor7Days() {
-        List<PoplarTagResDto> popularTagResDtoListByRedis = popularTagService.getLimitPopularTagListFor7Days();
+        List<PopularTagResDto> popularTagResDtoListByRedis = hashtagCategoryService.getLimitPopularTagListFor7Days();
 
         //redis 조회 시 빈값 체크
         //빈값이라면 전체 데이터 중 상위 8개 인기태그 불러오기
         if (CollectionUtils.isEmpty(popularTagResDtoListByRedis)){
-            List<PoplarTagResDto> popularTagResDtoList = popularTagService.getLimitPopularTagList();
+            List<PopularTagResDto> popularTagResDtoList = hashtagCategoryService.getLimitPopularTagList();
 
             return new ResponseEntity<>(new CMRespDto<>(
                     1, "전체 기간 상위 8개 인기태그 리스트 불러오기 성공", popularTagResDtoList), HttpStatus.OK);
