+++ src/main/java/com/web/doitcommit/dto/hashtagCategory/PopularTagResDto.java
@@ -1,10 +1,10 @@
-package com.web.doitcommit.dto.popularTag;
+package com.web.doitcommit.dto.hashtagCategory;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 @Data
-public class PoplarTagResDto {
+public class PopularTagResDto {
 
     @Schema(description = "태그 고유값")
     private Long tagId;
@@ -15,7 +15,7 @@ public class PoplarTagResDto {
     @Schema(description = "태그 수")
     private int count;
 
-    public PoplarTagResDto(Long tagId, String tagName, int count) {
+    public PopularTagResDto(Long tagId, String tagName, int count) {
         this.tagId = tagId;
         this.tagName = tagName;
         this.count = count;
