+++ src/main/java/com/web/doitcommit/service/hashtagCategory/HashtagCategoryService.java
@@ -1,7 +1,9 @@
-package com.web.doitcommit.service.popularTag;
+package com.web.doitcommit.service.hashtagCategory;
 
+import com.web.doitcommit.domain.hashtagCategory.HashtagCategory;
 import com.web.doitcommit.domain.hashtagCategory.HashtagCategoryRepository;
-import com.web.doitcommit.dto.popularTag.PoplarTagResDto;
+import com.web.doitcommit.dto.hashtagCategory.HashtagCategoryResDto;
+import com.web.doitcommit.dto.hashtagCategory.PopularTagResDto;
 import com.web.doitcommit.redis.RedisService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,10 +19,19 @@ import java.util.stream.Collectors;
 @Transactional
 @RequiredArgsConstructor
 @Service
-public class PopularTagService {
+public class HashtagCategoryService {
 
     private final HashtagCategoryRepository hashtagCategoryRepository;
     private final RedisService redisService;
+
+    /**
+     * 해시태그 목록 조회
+     */
+    @Transactional(readOnly = true)
+    public List<HashtagCategoryResDto> getBoardTagList() {
+        List<HashtagCategory> result = hashtagCategoryRepository.findAll();
+        return result.stream().map(tag -> new HashtagCategoryResDto(tag)).collect(Collectors.toList());
+    }
 
     /**
      * Redis 인기태그 갱신
@@ -46,11 +57,11 @@ public class PopularTagService {
      * 상단 8개 인기태그 리스트 - redis 조회
      */
     @Transactional(readOnly = true)
-    public List<PoplarTagResDto> getLimitPopularTagListFor7Days(){
+    public List<PopularTagResDto> getLimitPopularTagListFor7Days(){
 
         List<Object[]> result = redisService.getValues(LocalDate.now().toString());
 
-        return result.stream().map(arr -> new PoplarTagResDto(
+        return result.stream().map(arr -> new PopularTagResDto(
                 (Long) arr[0], (String) arr[1], ((Long) arr[2]).intValue())).collect(Collectors.toList());
     }
 
@@ -58,11 +69,11 @@ public class PopularTagService {
      * 상단 8개 인기태그 리스트 조회 - db 조회
      */
     @Transactional(readOnly = true)
-    public List<PoplarTagResDto> getLimitPopularTagList(){
+    public List<PopularTagResDto> getLimitPopularTagList(){
 
         List<Object[]> result = hashtagCategoryRepository.getLimitPopularTagList();
 
-        return result.stream().map(arr -> new PoplarTagResDto(
+        return result.stream().map(arr -> new PopularTagResDto(
                 (Long) arr[0], (String) arr[1], ((Long) arr[2]).intValue())).collect(Collectors.toList());
     }
 
@@ -70,13 +81,13 @@ public class PopularTagService {
      * 지정된 기간동안 {tagId, tagName, count} 모든 태그 리스트
      */
     @Transactional(readOnly = true)
-    public List<PoplarTagResDto> getAllPopularTagListForPeriod(int period){
+    public List<PopularTagResDto> getAllPopularTagListForPeriod(int period){
 
         List<Object[]> result = hashtagCategoryRepository.getAllPopularTagListForPeriod(period);
 
         System.out.println(result.size());
 
-        return result.stream().map(arr -> new PoplarTagResDto(
+        return result.stream().map(arr -> new PopularTagResDto(
                 (Long) arr[0], (String) arr[1], ((Long) arr[2]).intValue())).collect(Collectors.toList());
     }
 }
